Specify some project dependencies
Specifcally:

* PHP 5.4+ is required for PHP's traits support (this could be reduced
  without PSR logging, but PHP 5.3+ would still be required for PHP's
  namespace support.
* the PHP shmop extension is required (naturally, this whole project is
  based on the use of shmop functions).

This change also corrects the spelling of capitalisation of the
MetricsLoggerTest::setUp function.  Though PHPUnit is lenient, and the
change has no affect, it is now more aligned with the official PHPUnit
documentation.

Finally, this change also drops the strict_types declaration to improve
compatibility with PHP 5.x and HHVM (strict_types is currently only
supported by PHP 7, and we're not applying that feature yet anyway).